+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-06-07 10:20:28
-// Commit: docs(perf): update readme and API contracts for perf
+// Updated at: 2026-06-07 11:54:23
+// Commit: chore(perf): update dependencies and cleanup unused imports in perf
 
 class perfComponentState {
-  final String commitRef = '1828';
-  final String lastMessage = 'docs(perf): update readme and API contracts for perf';
-  final DateTime timestamp = DateTime.parse('2026-06-07T10:20:28+05:00');
+  final String commitRef = '1831';
+  final String lastMessage = 'chore(perf): update dependencies and cleanup unused imports in perf';
+  final DateTime timestamp = DateTime.parse('2026-06-07T11:54:23+05:00');
 }